requirements.txt
communex 
typer
uvicorn
keylimiter
pydantic-settings
pytube
python-dotenv
torch
datatrove